refactor: split oversized files and enforce line-length limits
Major restructuring to bring all Python files under 400-line cap (source) /
600-line cap (tests), enforced by new check-file-length pre-commit hook.

Handler refactoring:
- config_disclosure_handler.py: 1354→200 lines; extract response builders
  into handlers/config_responses/ package (php_configs, server_configs,
  yaml_json_configs, security_configs)
- routes.py: split into handlers/routes/ package with per-service files
  (routes_wordpress, routes_phpmyadmin, etc.)
- base_handler.py: 692→330 lines; remove dead code, consolidate methods
- http_handler.py: 618→273 lines; extract SSH/non-HTTP handlers, simplify
  routing logic
- webdav_handler.py: 455→216 lines; consolidate response builders
- bot_profile.py: 409→327 lines; extract pattern constants to patterns.py

Client refactoring:
- client.py: 510→221 lines; split into report_sender, server_factory,
  ssh_creds modules

Config/DB refactoring:
- config.py: 430→369 lines; fix _find_config_file() to remove in-package
  fallback that masked missing secrets; fix _resolve() None default handling
- storage.py: 439→289 lines; extract SQL builder

Test restructuring:
- test_client.py → test/test_client/ (multiple files)
- test_config.py → test/test_config/ (multiple files)
- test_integration.py → test/test_integration/ (multiple files)
- test_storage.py → test/test_storage/ (multiple files)
- test_http_handler.py: fix mock targets (_send_report → _enrich_and_send)

New infrastructure:
- scripts/check_file_length.py: enforces 400-line source / 600-line test caps
- handlers/patterns.py: extracted SQLI/LFI/RCE/ENUM/SCANNER/TOOL patterns
- handlers/config_responses/__init__.py: re-exports all response builders

Bug fixes:
- config validation: _find_config_file() no longer falls back to
  settings.toml.example (which had hivepass hardcoded), fixing the test
  that expected SystemExit(1) on missing secrets
- webdav_handler: restore missing _build_http_response() method
- http_handler tests: update mock targets to actual method names

All 438 tests pass at 75% coverage. All files under line limits.
ruff check passes clean.
