+++ manyfaced/handlers/config_responses/php_configs.py
@@ -1,0 +1,264 @@
+"""PHP configuration file responses for ConfigDisclosureHandler."""
+
+from __future__ import annotations
+
+
+def fake_wp_config_php() -> str:
+    """Fake wp-config.php with realistic but fake credentials."""
+    return r"""<?php
+/**
+ * Custom WordPress configurations on "wp-config.php" file.
+ *
+ * This file has the following configurations: MySQL settings, Table Prefix, Secret Keys, WordPress Language, ABSPATH and more.
+ * For more information visit {@link https://codex.wordpress.org/Editing_wp-config.php Editing wp-config.php} Codex page.
+ * Created using {@link http://generatewp.com/wp-config/ wp-config.php File Generator} on GenerateWP.com.
+ *
+ * @package WordPress
+ * @generator GenerateWP.com
+ */
+
+/* MySQL settings */
+define( 'DB_NAME',     'wpdb' );
+define( 'DB_USER',     'root' );
+define( 'DB_PASSWORD', 'jadolbaeb' );
+define( 'DB_HOST',     'localhost' );
+define( 'DB_CHARSET',  'utf8mb4' );
+define( 'DB_COLLATE',  '' );
+
+/* MySQL database table prefix. */
+$table_prefix = 'wp_';
+
+/* Authentication Unique Keys and Salts. */
+define('AUTH_KEY',         'Xm3Kp9vN2lR8wQ7jH5tY6uI1oP0aS4dFgBcVnMxZlKjHgFdSaWqWeRtYuIoP');
+define('SECURE_AUTH_KEY',  'QwErTyUiOpAsDfGhJkLzXcVbNmWqWeRtYuIoPaSdFgHjKlZxCvBnM');
+define('LOGGED_IN_KEY',    'ZxCvBnMqWeRtYuIoPaSdFgHjKlZxQwErTyUiOpAsDfGhJkLzXcVbNm');
+define('NONCE_KEY',        'LkMjNhBgYtUeWrOpIaDsFgHjKlZxQwErTyUiOpAsDfGhJkLzXcVbNm');
+define('AUTH_SALT',        'BnMqWeRtYuIoPaSdFgHjKlZxQwErTyUiOpAsDfGhJkLzXcVbNmQwEr');
+define('SECURE_AUTH_SALT', 'TyUiOpAsDfGhJkLzXcVbNmQwErTyUiOpAsDfGhJkLzXcVbNmQwErTy');
+define('LOGGED_IN_SALT',   'UiOpAsDfGhJkLzXcVbNmQwErTyUiOpAsDfGhJkLzXcVbNmQwErTyUi');
+define('NONCE_SALT',       'OpAsDfGhJkLzXcVbNmQwErTyUiOpAsDfGhJkLzXcVbNmQwErTyUiOp');
+
+/* WordPress Localized Language. */
+define('WPLANG', '');
+define('WP_LANG_DIR', WP_CONTENT_DIR . '/languages');
+
+/* WordPress Database Table ID. */
+$wpdb->insert_id = 0;
+
+/* That's all, stop editing! Happy blogging. */
+
+/** Absolute path to the WordPress directory. */
+if ( !defined('ABSPATH') )
+	define('ABSPATH', dirname(__FILE__) . '/');
+
+/** Sets up WordPress vars and included files. */
+require_once(ABSPATH . 'wp-settings.php');"""
+
+
+def fake_xmlrpc_php() -> str:
+    """Fake xmlrpc.php response."""
+    return r"""<?xml version="1.0" encoding="UTF-8"?><methodResponse><params><param><value><struct><member><name>faultCode</name><value><int>-32700</int></value></member><member><name>faultString</name><value>Parse error: syntax error, unexpected end of file in /var/www/html/xmlrpc.php on line 145 - XML-RPC server does not support this method</value></member></struct></value></param></params></methodResponse>"""
+
+
+def fake_config_php() -> str:
+    """Fake Drupal config.php."""
+    return r"""<?php
+/**
+ * @file
+ * Drupal site-specific configuration file.
+ */
+
+$databases = array (
+  'default' => 
+  array (
+    'default' => 
+    array (
+      'database' => 'drupaldb',
+      'username' => 'root',
+      'password' => 'drup4lp@ss!',
+      'host' => 'localhost',
+      'port' => '',
+      'driver' => 'mysql',
+      'prefix' => '',
+    ),
+  ),
+);
+
+$conf['site_name'] = 'My Drupal Site';
+$conf['file_public_path'] = 'sites/default/files';
+$conf['cache_enable'] = 1;"""
+
+
+def fake_phpinfo_php() -> str:
+    """Fake phpinfo.php output."""
+    return r"""<!DOCTYPE html>
+<html lang="en">
+<head><meta charset="UTF-8"><title>phpinfo()</title></head>
+<body>
+<h1>PHP Version 8.2.15</h1>
+<p>This page displays phpinfo() output for debugging purposes.</p>
+<table>
+<tr><td>System</td><td>Linux webserver 5.15.0-91-generic #101-Ubuntu SMP x86_64</td></tr>
+<tr><td>Server API</td><td>Apache 2.0 Handler</td></tr>
+<tr><td>Configuration File (php.ini) Path</td><td>/etc/php/8.2/apache2</td></tr>
+<tr><td>Loaded Configuration File</td><td>/etc/php/8.2/apache2/php.ini</td></tr>
+<tr><td>Scan this dir for additional .ini files</td><td>/etc/php/8.2/apache2/conf.d</td></tr>
+<tr><td>Additional .ini files parsed</td><td>(none)</td></tr>
+</table>
+<h2>Registered PHP Streams</h2>
+<p>https, ftps, php, file, glob, data, http, ftp, phar, zip</p>
+<h2>Registered Stream Socket Transports</h2>
+<p>tcp, udp, unix, udg, ssl, tls, tlsv1.0, tlsv1.1, tlsv1.2, tlsv1.3</p>
+</body></html>"""
+
+
+def fake_php_ini() -> str:
+    """Fake php.ini content."""
+    return r"""; PHP Settings for Production Server
+; Generated by manyfaced honeypot
+
+[PHP]
+engine = On
+short_open_tag = Off
+precision = 14
+output_buffering = 4096
+zlib.output_compression = Off
+implicit_flush = Off
+unserialize_callback_func =
+serialize_precision = -1
+disable_functions = exec,passthru,shell_exec,system
+disable_classes = 
+zend.enable_gc = On
+expose_php = On
+max_execution_time = 30
+max_input_time = 60
+memory_limit = 256M
+error_reporting = E_ALL & ~E_DEPRECATED & ~E_STRICT
+display_errors = Off
+display_startup_errors = Off
+log_errors = On
+log_errors_max_len = 1024
+ignore_repeated_errors = Off
+ignore_repeated_source = Off
+report_memleaks = On
+variables_order = "GPCS"
+request_order = "GP"
+register_globals = Off
+post_max_size = 50M
+auto_prepend_file = 
+auto_append_file = 
+default_mimetype = "text/html"
+default_charset = "UTF-8"
+
+[Date]
+date.timezone = UTC
+
+[MySQLi]
+mysqli.max_persistent = -1
+mysqli.allow_local_infile = On
+
+[Session]
+session.save_handler = files
+session.use_strict_mode = 0
+session.use_cookies = 1
+session.use_only_cookies = 1
+session.name = PHPSESSID
+session.auto_start = 0
+session.cookie_lifetime = 0
+session.cookie_path = /
+session.cookie_domain = 
+session.cookie_httponly = On
+session.cookie_secure = Off
+session.serialize_handler = php
+session.gc_probability = 1
+session.gc_divisor = 1000
+session.referer_check = 
+
+[mail function]
+SMTP = localhost
+smtp_port = 25
+mail.add_x_header = On"""
+
+
+def fake_settings_py() -> str:
+    """Fake Django settings.py."""
+    return r"""# Django settings configuration
+# Generated: 2024-01-15
+
+import os
+from pathlib import Path
+
+BASE_DIR = Path(__file__).resolve().parent.parent
+
+SECRET_KEY='django...7890'
+
+DEBUG = True
+
+ALLOWED_HOSTS = ['localhost', '127.0.0.1', 'example.com', '*.example.com']
+
+INSTALLED_APPS = [
+    'django.contrib.admin',
+    'django.contrib.auth',
+    'django.contrib.contenttypes',
+    'django.contrib.sessions',
+    'django.contrib.messages',
+    'django.contrib.staticfiles',
+    'django.contrib.sites',
+    'django.contrib.admindocs',
+    'django.contrib.sitemaps',
+    'django.contrib.humanize',
+]
+
+MIDDLEWARE = [
+    'django.middleware.security.SecurityMiddleware',
+    'django.contrib.sessions.middleware.SessionMiddleware',
+    'django.middleware.common.CommonMiddleware',
+    'django.middleware.csrf.CsrfViewMiddleware',
+    'django.contrib.auth.middleware.AuthenticationMiddleware',
+    'django.contrib.messages.middleware.MessageMiddleware',
+    'django.middleware.clickjacking.XFrameOptionsMiddleware',
+]
+
+ROOT_URLCONF = 'myproject.urls'
+
+TEMPLATES = [
+    {
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        'DIRS': [BASE_DIR / 'templates'],
+        'APP_DIRS': True,
+        'OPTIONS': {
+            'context_processors': [
+                'django.template.context_processors.debug',
+                'django.template.context_processors.request',
+                'django.contrib.auth.context_processors.auth',
+                'django.contrib.messages.context_processors.messages',
+            ],
+        },
+    },
+]
+
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.mysql',
+        'NAME': 'myproject_db',
+        'USER': 'root',
+        'PASSWORD': 'Djang0DBP@ss!',
+        'HOST': 'localhost',
+        'PORT': '3306',
+    }
+}
+
+AUTH_PASSWORD_VALIDATORS = [
+    {'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator'},
+    {'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator'},
+    {'NAME': 'django.contrib.auth.password_validation.CommonPasswordValidator'},
+    {'NAME': 'django.contrib.auth.password_validation.NumericPasswordValidator'},
+]
+
+LANGUAGE_CODE = 'en-us'
+TIME_ZONE = 'UTC'
+USE_I18N = True
+USE_TZ = True
+
+STATIC_URL = '/static/'
+DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'"""
