+++ manyfaced/handlers/http_handler.py
@@ -1,21 +1,14 @@
 """HTTPHandler – handles raw HTTP requests from bots.
 
-This handler receives raw HTTP data from connecting bots, routes requests
-to the appropriate service handler via an explicit Router, generates
-realistic honeypot responses, and spawns processes to send reports to
-the server.
-
-The router walks an ordered route table and returns on first match — no
-more response concatenation across multiple handlers.
+Routes to service-specific handlers via Router, generates honeypot responses,
+and queues reports for sending to the server. Response content is delegated
+to protocol_responses module; report queue management to report_queue module.
 """
 
 from __future__ import annotations
 
 import os
-import queue as _queue
 import random
-import threading
-from concurrent.futures import ThreadPoolExecutor
 from datetime import datetime, timezone
 
 from manyfaced.common.bearstorage import BearStorage
@@ -32,93 +25,28 @@ from manyfaced.common.status import (
     UNKNOWN_REDIS,
     UNKNOWN_TLS,
 )
-
-from .config_disclosure_handler import ConfigDisclosureHandler
-from .cpanel_handler import CPanelHandler
-from .drupal_handler import DrupalHandler
-from .generic_handler import GenericHandler
-from .jenkins_handler import JenkinsHandler
-from .phpmyadmin_handler import PhpMyAdminHandler
-from .router import Router
-from .tomcat_handler import TomcatHandler
-from .wordpress_handler import WordPressHandler
+from manyfaced.handlers.protocol_responses import (
+    fallback_response,
+    fake_ssh_banner,
+    ftp_banners,
+    non_http_response,
+)
+from manyfaced.handlers.report_queue import _get_report_queue, shutdown_report_executor
 
 logger = get_logger(__name__)
 
 
 # Singleton router – initialized on first use
-_router: Router | None = None
-
-# Bounded work queue for sending reports (replaces per-request subprocess spawning).
-# Uses a queue.Queue with maxsize for backpressure: when the queue is full,
-# put() blocks until space is available.
-# Worker threads pull from the queue and execute the send_report function.
-_report_queue: _queue.Queue | None = None
-_report_queue_lock = threading.Lock()
-_report_queue_alive: bool = False  # Public flag for aliveness tracking
-_report_workers: list[threading.Thread] = []
-_report_workers_lock = threading.Lock()
-
-MAX_REPORT_THREADS = 10
+_router: object | None = None
 
 
-def _report_worker():
-    """Worker thread that processes items from the report queue."""
-    q = _get_report_queue()
-    while _report_queue_alive:
-        try:
-            fn, args = q.get(timeout=1)
-            try:
-                fn(*args)
-            except Exception:
-                logger.exception('Report worker error')
-            finally:
-                q.task_done()
-        except _queue.Empty:
-            continue
-
-
-def _get_report_queue() -> _queue.Queue:
-    """Get or create the module-level report work queue (singleton).
-
-    Uses a bounded queue (maxsize=MAX_REPORT_THREADS*10) to provide
-    backpressure: when the queue is full, put() blocks until space is available.
-    """
-    global _report_queue, _report_queue_alive
-    if _report_queue is None:
-        with _report_queue_lock:
-            if _report_queue is None:
-                _report_queue = _queue.Queue(maxsize=MAX_REPORT_THREADS * 10)
-                _report_queue_alive = True
-                with _report_workers_lock:
-                    for _ in range(MAX_REPORT_THREADS):
-                        t = threading.Thread(
-                            target=_report_worker,
-                            daemon=True,
-                            name='report_worker',
-                        )
-                        t.start()
-                        _report_workers.append(t)
-    return _report_queue
-
-
-def shutdown_report_executor():
-    """Gracefully shut down the report work queue and workers."""
-    global _report_queue, _report_queue_alive, _report_workers
-    if _report_queue is not None and _report_queue_alive:
-        _report_queue_alive = False
-        # Wait for queue to drain
-        _report_queue.join()
-        _report_queue = None
-        with _report_workers_lock:
-            _report_workers.clear()
-
-
-def _get_router() -> Router:
+def _get_router():
     """Get or create the module-level router (singleton)."""
     global _router
     if _router is None:
-        from .routes import ROUTES  # noqa: F811
+        from manyfaced.handlers.routes import ROUTES  # noqa: F811
+
+        from manyfaced.handlers.router import Router
 
         _router = Router(ROUTES)
         logger.info('Router initialized with %d routes', len(_router.routes))
@@ -130,33 +58,15 @@ class HTTPHandler:
 
     Unlike the server-side BaseHandler, this does NOT decrypt or parse JSON.
     It receives raw HTTP data, routes to the appropriate handler, generates
-    a honeypot response, and spawns a process to send the report to the server.
+    a honeypot response, and queues reports for sending to the server.
     """
 
     def __init__(self, args, update_event):
-        """Initialize the HTTP handler.
-
-        Args:
-            args: CLI arguments namespace
-            update_event: Event to signal shutdown
-        """
         self.args = args
         self.update_event = update_event
 
     def handle_request(self, message: str, bot_ip: str = '127.0.0.1'):
-        """Handle a raw HTTP request from a bot.
-
-        Detects protocol before parsing and handles non-HTTP probes
-        (SSH, FTP, etc.) with appropriate responses and reports.
-
-        Args:
-            message: Raw HTTP request string from the bot.
-            bot_ip: IP address of the connecting bot.
-
-        Returns:
-            The honeypot response data (HTTP response string or bytes).
-        """
-        # Detect protocol before attempting HTTP parsing
+        """Handle a raw HTTP request from a bot."""
         raw_bytes = message.encode('utf-8') if isinstance(message, str) else message
         if not raw_bytes:
             return self._handle_empty_connection(bot_ip)
@@ -179,122 +89,29 @@ class HTTPHandler:
         # Parse the raw HTTP request
         try:
             parsed = HTTPRequest(message)
-            # If parsing failed (path is None), create a minimal valid request
-            path_val = getattr(parsed, 'path', None)
-            if path_val is None:
+            if getattr(parsed, 'path', None) is None:
                 logger.debug('HTTPRequest failed to parse path, using fallback for %s', bot_ip)
-                fallback = 'GET / HTTP/1.1\r\nHost: localhost\r\nUser-Agent: Unknown\r\n\r\n'
-                parsed = HTTPRequest(fallback)
+                parsed = HTTPRequest('GET / HTTP/1.1\r\nHost: localhost\r\nUser-Agent: Unknown\r\n\r\n')
         except Exception as e:
             logger.debug('Failed to parse HTTP request: %s, using fallback for %s', e, bot_ip)
-            fallback = 'GET / HTTP/1.1\r\nHost: localhost\r\nUser-Agent: Unknown\r\n\r\n'
-            parsed = HTTPRequest(fallback)
+            parsed = HTTPRequest('GET / HTTP/1.1\r\nHost: localhost\r\nUser-Agent: Unknown\r\n\r\n')
 
-        # Build the data dict that process_request expects
-        # Use original message as raw_request; if empty, use fallback to ensure capture
-        raw_for_report = (
-            message
-            if message
-            else ('GET / HTTP/1.1\r\nHost: localhost\r\nUser-Agent: Unknown\r\n\r\n')
-        )
+        raw_for_report = message if message else 'GET / HTTP/1.1\r\nHost: localhost\r\nUser-Agent: Unknown\r\n\r\n'
         data = {
             'ip': bot_ip,
             'raw_request': raw_for_report,
             'parsed_request': parsed,
         }
-
         return self.process_request(data)
 
-    def _send_report(
-        self, bot_ip: str, raw_request: str, parsed, detected: int, protocol: str = None
-    ):
-        """Send a report to the server for a bot interaction.
-
-        Args:
-            bot_ip: Bot IP address.
-            raw_request: Raw request data.
-            parsed: Parsed request object.
-            detected: Detection ID.
-            protocol: Detected protocol name (for non-HTTP).
-        """
-        from manyfaced.client.client import send_report
-
-        request_time = str(datetime.now(timezone.utc).strftime('%Y-%m-%d %H:%M:%S.%f'))
-        server_host = getattr(self.args, 'server_host', '127.0.0.1')
-        server_port = getattr(self.args, 'server', None)
-
-        if server_port is None:
-            return
-
-        bs = BearStorage(
-            bot_ip,
-            raw_request,
-            request_time,
-            parsed,
-            detected,
-            settings.HIVELOGIN,
-        )
-
-        q = _get_report_queue()
-        q.put(
-            (
-                send_report,
-                (bs, bot_ip, settings.HIVEPASS, server_host, server_port, settings.HIVELOGIN),
-            )
-        )
-
-    def _send_report_enriched(self, bs: BearStorage, bot_ip: str) -> None:
-        """Send an already-enriched report to the server.
-
-        Used by SSH and non-HTTP handlers that perform their own enrichment
-        before calling this method (matching the HTTP path ordering in process_request).
-
-        Args:
-            bs: BearStorage instance with dns_name, country, continent already populated.
-            bot_ip: Bot IP address.
-        """
-        from manyfaced.client.client import send_report
-
-        server_host = getattr(self.args, 'server_host', '127.0.0.1')
-        server_port = getattr(self.args, 'server', None)
-
-        if server_port is None:
-            return
-
-        q = _get_report_queue()
-        q.put(
-            (
-                send_report,
-                (bs, bot_ip, settings.HIVEPASS, server_host, server_port, settings.HIVELOGIN),
-            )
-        )
-
     def _handle_ssh_probe(self, bot_ip: str, protocol_info: dict) -> bytes:
-        """Handle an SSH probe by responding with a fake SSH banner and capturing credentials.
-
-        Args:
-            bot_ip: IP address of the connecting bot.
-            protocol_info: Dict with SSH protocol details.
-
-        Returns:
-            Fake SSH banner response bytes.
-        """
-        # Extract client version if available
+        """Handle an SSH probe by responding with a fake SSH banner."""
         client = protocol_info.get('client', '')
         version = protocol_info.get('version', '')
-
-        # Generate a realistic-looking SSH banner with varying OpenSSH versions
-        banner_versions = [
-            'SSH-2.0-OpenSSH_9.6',
-            'SSH-2.0-OpenSSH_8.9',
-            'SSH-2.0-OpenSSH_7.9',
-            'SSH-2.0-libssh2_1.10.0',
-        ]
-        banner = random.choice(banner_versions) + '\r\n'
+        banner = fake_ssh_banner() + '\r\n'
 
         logger.debug('Sent SSH banner to %s (client=%s)', bot_ip, client)
 
-        # Enrich with DNS and geo data (same pattern as HTTP path in process_request)
         class _ParsedSSH:
             command = 'SSH'
             path = '/'
@@ -303,100 +120,24 @@ class HTTPHandler:
             request_version = version or 'SSH-2.0'
 
         bs = BearStorage(
-            bot_ip,
-            protocol_info.get('raw', 'SSH-2.0-PUTTY'),
+            bot_ip, protocol_info.get('raw', 'SSH-2.0-PUTTY'),
             str(datetime.now(timezone.utc).strftime('%Y-%m-%d %H:%M:%S.%f')),
-            _ParsedSSH(),
-            SSH_CLIENT,
-            settings.HIVELOGIN,
+            _ParsedSSH(), SSH_CLIENT, settings.HIVELOGIN,
         )
-
-        # Resolve reverse-DNS off the hot path (with short timeout)
-        try:
-            bs.dns_name = bs.resolve_dns_name(bot_ip, timeout=1.0)
-        except Exception:
-            logger.debug('DNS resolution failed for %s', bot_ip)
-
-        # Resolve geolocation off the hot path (with short timeout)
-        try:
-            bs.resolve_geo(bot_ip, timeout=2.0)
-        except Exception:
-            logger.debug('Geo resolution failed for %s', bot_ip)
-
-        # Send enriched report for SSH probe
-        self._send_report_enriched(bs, bot_ip)
+        self._enrich_and_send(bs, bot_ip)
         return banner.encode('utf-8')
 
     def _handle_non_http_probe(self, bot_ip: str, protocol: str, protocol_info: dict) -> bytes:
-        """Handle non-HTTP protocol probes.
+        """Handle non-HTTP protocol probes."""
+        detected_id = {
+            'tls': UNKNOWN_TLS, 'dns': UNKNOWN_DNS,
+            'mongodb': UNKNOWN_MONGODB, 'redis': UNKNOWN_REDIS,
+        }.get(protocol, UNKNOWN_NON_HTTP)
 
-        Routes each recognized non-HTTP protocol to its own detected_id so
-        future work can build per-protocol response handlers. Classification
-        only — all protocols still get the generic monster page back.
-
-        Args:
-            bot_ip: IP address of the connecting bot.
-            protocol: Detected protocol name.
-            protocol_info: Dict with protocol details.
-
-        Returns:
-            Appropriate response bytes for the protocol.
-        """
-        # Per-protocol detected_id mapping
-        if protocol == 'tls':
-            detected_id = UNKNOWN_TLS
-        elif protocol == 'dns':
-            detected_id = UNKNOWN_DNS
-        elif protocol == 'mongodb':
-            detected_id = UNKNOWN_MONGODB
-        elif protocol == 'redis':
-            detected_id = UNKNOWN_REDIS
-        else:
-            # Existing protocols (ftp, telnet, rdp, vnc, smtp, pop3, imap)
-            detected_id = UNKNOWN_NON_HTTP
-
-        response = None
-
-        # FTP: respond with a fake FTP banner
-        if protocol == 'ftp':
-            banners = [
-                '220 (vsFTPd 3.0.3)\r\n',
-                '220 Welcome to FTP service.\r\n',
-                '220 ProFTPD 1.3.5 Server\r\n',
-            ]
-            response = random.choice(banners).encode('utf-8')
-            detected_id = UNKNOWN_NON_HTTP
-
-        # Telnet: send null bytes (common telnet probe response)
-        elif protocol == 'telnet':
-            response = b'\xff\xfb\x01\xff\xfb\x03\xff\xfd\x1f'
-            detected_id = UNKNOWN_NON_HTTP
-
-        # RDP: send a negative response
-        elif protocol == 'rdp':
-            response = b'\x03\x00\x00\x1f\x0e\xe0\x00\x00\x18\x00\x01\xc1\x00\x00\x00'
-            detected_id = UNKNOWN_NON_HTTP
-
-        # VNC: respond with version string
-        elif protocol == 'vnc':
-            response = b'RFB 003.003\r\n'
-            detected_id = UNKNOWN_NON_HTTP
-
-        # SMTP/POP3/IMAP: send a fake greeting
-        elif protocol in ('smtp', 'pop3', 'imap'):
-            greetings = {
-                'smtp': '220 manyfaced-honeypot ESMTP\r\n',
-                'pop3': '+OK manyfaced-honeypot POP3 ready\r\n',
-                'imap': '* OK manyfaced-honeypot IMAP4 ready\r\n',
-            }
-            response = greetings[protocol].encode('utf-8')
-            detected_id = UNKNOWN_NON_HTTP
-
-        else:
-            # Unknown protocol: just close (empty response)
+        response = non_http_response(protocol)
+        if response is None:
             response = b''
 
-        # Enrich with DNS and geo data for all recognized non-HTTP protocols
         class _ParsedNonHTTP:
             command = protocol.upper()
             path = '/'
@@ -405,44 +146,16 @@ class HTTPHandler:
             user_agent = protocol_info.get('client', protocol)
 
         bs = BearStorage(
-            bot_ip,
-            protocol_info.get('raw', ''),
+            bot_ip, protocol_info.get('raw', ''),
             str(datetime.now(timezone.utc).strftime('%Y-%m-%d %H:%M:%S.%f')),
-            _ParsedNonHTTP(),
-            detected_id,
-            settings.HIVELOGIN,
+            _ParsedNonHTTP(), detected_id, settings.HIVELOGIN,
         )
-
-        # Resolve reverse-DNS off the hot path (with short timeout)
-        try:
-            bs.dns_name = bs.resolve_dns_name(bot_ip, timeout=1.0)
-        except Exception:
-            logger.debug('DNS resolution failed for %s', bot_ip)
-
-        # Resolve geolocation off the hot path (with short timeout)
-        try:
-            bs.resolve_geo(bot_ip, timeout=2.0)
-        except Exception:
-            logger.debug('Geo resolution failed for %s', bot_ip)
-
-        # Send enriched report for non-HTTP probe
-        self._send_report_enriched(bs, bot_ip)
-
+        self._enrich_and_send(bs, bot_ip)
         return response
 
     def process_request(self, data):
-        """Process an incoming HTTP request.
-
-        Routes to the appropriate handler, generates response, and
-        spawns a process to send the report to the server.
-
-        Args:
-            data: Dict with 'ip', 'raw_request', 'parsed_request'
-
-        Returns:
-            The honeypot response bytes.
-        """
-        from manyfaced.client.client import send_report
+        """Process an incoming HTTP request."""
+        from manyfaced.client.client import send_report  # noqa: PLC0415
 
         bot_ip = data['ip']
         raw_request = data['raw_request']
@@ -451,7 +164,6 @@ class HTTPHandler:
 
         logger.info('Incoming request from %s at %s', bot_ip, request_time)
 
-        # Extract headers from parsed request
         headers = {}
         if hasattr(parsed, 'headers') and parsed.headers:
             try:
@@ -459,102 +171,25 @@ class HTTPHandler:
             except Exception:
                 logger.debug('Failed to parse request headers for %s', bot_ip)
 
-        # Route through the router (first match wins)
         router = _get_router()
         path = getattr(parsed, 'path', '/')
-
-        # Try router dispatch
         result = router.dispatch(path, raw_request, bot_ip, headers or {})
 
         if result is not None:
             output_data, detected = result
-            logger.debug(
-                'Router returned response for %s (detected=%s)',
-                bot_ip,
-                detected,
-            )
-
-            # Record the full interaction in the dialogue for the matched handler
-            request_data = {
-                'path': path,
-                'method': self._extract_method(raw_request),
-                'raw': raw_request,
-                'headers': headers,
-            }
-            # The router dispatch already called generate_response() which records
-            # the interaction via profile.record_interaction(). No need to iterate.
         else:
-            output_data, detected = self._fallback_response(path), 1
+            output_data, detected = fallback_response(path), 1
 
-        logger.debug(
-            'Generated response for %s, detected=%s, size=%d',
-            bot_ip,
-            detected,
-            len(output_data),
-        )
+        logger.debug('Generated response for %s, detected=%s, size=%d', bot_ip, detected, len(output_data))
 
-        # Create BearStorage for reporting
-        logger.debug('Creating BearStorage for %s', bot_ip)
         bs = BearStorage(
-            bot_ip,
-            raw_request,
-            request_time,
-            parsed,
-            detected,
-            settings.HIVELOGIN,
+            bot_ip, raw_request, request_time, parsed, detected, settings.HIVELOGIN,
         )
-        logger.debug('BearStorage created for %s', bot_ip)
-
-        # Resolve reverse-DNS off the hot path (with short timeout)
-        try:
-            bs.dns_name = bs.resolve_dns_name(bot_ip, timeout=1.0)
-        except Exception:
-            logger.debug('DNS resolution failed for %s', bot_ip)
-
-        # Resolve geolocation off the hot path (with short timeout)
-        try:
-            bs.resolve_geo(bot_ip, timeout=2.0)
-        except Exception:
-            logger.debug('Geo resolution failed for %s', bot_ip)
-
-        # Determine server connection info for sending reports
-        server_host = getattr(self.args, 'server_host', '127.0.0.1')
-        server_port = getattr(self.args, 'server', None)
-
-        if server_port is not None:
-            # Use bounded queue for backpressure instead of per-request subprocess spawning.
-            # This prevents process explosion: the old code spawned 1 Process
-            # per bot request (200+ processes logged), causing file descriptor
-            # exhaustion and crashes. Worker threads pull from the queue.
-            q = _get_report_queue()
-            q.put(
-                (
-                    send_report,
-                    (bs, bot_ip, settings.HIVEPASS, server_host, server_port, settings.HIVELOGIN),
-                )
-            )
-        else:
-            logger.debug('No server port configured, skipping report for %s', bot_ip)
-
+        self._enrich_and_send(bs, bot_ip)
         return output_data
 
     def _handle_empty_connection(self, bot_ip: str) -> bytes:
-        """Handle a zero-byte connection (port scan with no data sent).
-
-        Constructs a minimal parsed object and sends an enriched report
-        directly with EMPTY_CONNECTION as detected_id. DNS reverse-lookup
-        and geo enrichment run inline (same pattern as SSH/non-HTTP probes),
-        not via _send_report. No user-agent extraction occurs since there
-        is no input to parse. The record stores empty request_raw, making
-        it distinct from real GET / probes.
-
-        Args:
-            bot_ip: IP address of the connecting bot.
-
-        Returns:
-            Honeypot response bytes (generic fallback).
-        """
-
+        """Handle a zero-byte connection (port scan with no data sent)."""
         class _ParsedEmpty:
             command = ''
             path = ''
@@ -563,31 +198,31 @@ class HTTPHandler:
             user_agent = ''
             request_version = ''
 
-        # Create BearStorage for reporting
         bs = BearStorage(
-            bot_ip,
-            '',  # empty raw_request — no data was received
-            str(datetime.now(timezone.utc).strftime('%Y-%m-%d %H:%M:%S.%f')),
-            _ParsedEmpty(),
-            EMPTY_CONNECTION,
-            settings.HIVELOGIN,
+            bot_ip, '', str(datetime.now(timezone.utc).strftime('%Y-%m-%d %H:%M:%S.%f')),
+            _ParsedEmpty(), EMPTY_CONNECTION, settings.HIVELOGIN,
         )
+        self._enrich_and_send(bs, bot_ip)
+        return fallback_response('')
 
-        # Resolve reverse-DNS off the hot path (with short timeout)
+    def _enrich_and_send(self, bs: BearStorage, bot_ip: str) -> None:
+        """Resolve DNS/geo and queue report for a BearStorage entry."""
         try:
             bs.dns_name = bs.resolve_dns_name(bot_ip, timeout=1.0)
         except Exception:
             logger.debug('DNS resolution failed for %s', bot_ip)
 
-        # Resolve geolocation off the hot path (with short timeout)
         try:
             bs.resolve_geo(bot_ip, timeout=2.0)
         except Exception:
             logger.debug('Geo resolution failed for %s', bot_ip)
 
-        # Send enriched report for empty connection
-        self._send_report_enriched(bs, bot_ip)
-        return self._fallback_response('')
+        server_host = getattr(self.args, 'server_host', '127.0.0.1')
+        server_port = getattr(self.args, 'server', None)
+        if server_port is not None:
+            q = _get_report_queue()
+            from manyfaced.client.client import send_report  # noqa: PLC0415
+            q.put((send_report, (bs, bot_ip, settings.HIVEPASS, server_host, server_port, settings.HIVELOGIN)))
 
     @staticmethod
     def _extract_method(raw_request: str) -> str:
@@ -596,23 +231,3 @@ class HTTPHandler:
         if parts and len(parts) >= 1:
             return parts[0].upper()
         return 'GET'
-
-    def _fallback_response(self, path: str) -> bytes:
-        """Fallback response for unmatched paths."""
-        now = datetime.now(timezone.utc).strftime('%a, %d %b %Y %H:%M:%S GMT')
-        body = f"""<!DOCTYPE html>
-<html><head><title>Server</title></head>
-<body><h1>Welcome to zlol's manyface!</h1>
-<p>Server: Apache/2.4.57 (Ubuntu)</p>
-<p>Path: {path}</p>
-</body></html>"""
-        response = (
-            f'HTTP/1.1 200 OK\r\n'
-            f'Server: Apache/2.4.57 (Ubuntu)\r\n'
-            f'Date: {now}\r\n'
-            f'Content-Type: text/html; charset=UTF-8\r\n'
-            f'Connection: close\r\n'
-            f'\r\n'
-            f'{body}'
-        )
-        return response.encode('iso-8859-1')
